+++ src/Advertiser/Dto/Input/InvalidInputException.php
@@ -27,5 +27,4 @@ use RuntimeException;
 
 class InvalidInputException extends RuntimeException
 {
-
 }