+++ src/Demand/Application/Exception/TransferMoneyException.php
@@ -27,5 +27,4 @@ use RuntimeException;
 
 class TransferMoneyException extends RuntimeException
 {
-
 }