+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Connecting your account with the cryptocurrency wallet
+- Log in to your account with the cryptocurrency wallet
+- Auto registration
+- Auto withdrawal
 
 ## [1.10.9] - 2021-12-09
 ### Added